Add language to everything and fetch data for bot

+++ packages/backend/src/index.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import { PrismaClient } from '@prisma/client';
-import { findHelpType, findLanguage, findOption } from './db/helpers';
+import { findHelpType, findLanguage, findOption, getUILanguages } from './db/helpers';
 
 const prisma = new PrismaClient();
 const port = 3000;
@@ -8,29 +8,43 @@ const port = 3000;
 const app = express();
 app.use(express.json());
 
+app.get('/languages', async (_, res) => {
+  try {
+    const uiLanguages = await getUILanguages(prisma);
+
+    res.status(200).json(uiLanguages);
+  } catch (e) {
+    res.status(500).send();
+  }
+});
+
 app.post('/register', async (req, res) => {
-  const { userId, chatId, language, option, helpType } = req.body;
+  const { userId, chatId, uiLanguage, option, helpType } = req.body;
+
+  try {
+    // @NOTE: Dev only, remove the user before storing to db, for easier testing
+    // @TODO: Remove later
+    await prisma.user.delete({ where: { telegramUserId: userId } });
+  } catch (e) {
+    console.log('First time user.');
+  }
 
   try {
     // @NOTE: These records are provided as strings from the bot.
     // Fetch them from their tables by name for now,
     // until we are able to provide them to the bot and populate with name and id
     const [languageRow, optionRow, helpTypeRow] = await Promise.all([
-      findLanguage(prisma, language),
+      findLanguage(prisma, uiLanguage),
       findOption(prisma, option),
       findHelpType(prisma, helpType)
     ]);
 
     if (languageRow?.id && optionRow?.id && helpTypeRow?.id) {
-      // @NOTE: Dev only, remove the user before storing to db, for easier testing
-      // @TODO: Remove later
-      await prisma.user.delete({ where: { userId } });
-
       const user = await prisma.user.create({
         data: {
-          userId,
           chatId,
-          languageId: languageRow.id,
+          telegramUserId: userId,
+          uiLanguageId: languageRow.id,
           optionId: optionRow.id,
           helpTypeId: helpTypeRow.id
         }
